+++ cli/src/git/indicator.rs
@@ -2,21 +2,24 @@ use maestro_git::{LocalChanges, PullResult, SyncKind, SyncResult};
 
 pub enum SyncIndicator {
     Clean,
-    LocalChanges,
     Error,
+    LocalChanges,
+    Noop,
 }
 
 impl From<&SyncResult> for SyncIndicator {
     fn from(sync_result: &SyncResult) -> Self {
-        let SyncKind::Pull(pull_result) = &sync_result.kind;
-        match pull_result {
-            PullResult::FastForward { .. } | PullResult::UpToDate => {
-                match sync_result.state.changes {
-                    LocalChanges::Clean => SyncIndicator::Clean,
-                    LocalChanges::Present { .. } => SyncIndicator::LocalChanges,
+        match &sync_result.kind {
+            SyncKind::Pull(pull_result) => match pull_result {
+                PullResult::FastForward { .. } | PullResult::UpToDate => {
+                    match sync_result.state.changes {
+                        LocalChanges::Clean => SyncIndicator::Clean,
+                        LocalChanges::Present { .. } => SyncIndicator::LocalChanges,
+                    }
                 }
-            }
-            _ => SyncIndicator::Error,
+                _ => SyncIndicator::Error,
+            },
+            SyncKind::Skipped => SyncIndicator::Noop,
         }
     }
 }